backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-socketio==5.11.0
psycopg[binary]==3.2.4
google-generativeai==0.8.0
python-multipart==0.0.9
python-dotenv==1.0.1
aiofiles==24.1.0
